refactor: domain, view 분리
