+++ src/main/java/racingcar/domain/Car.java
@@ -1,26 +1,18 @@
-package racingcar;
+package racingcar.domain;
 
 public class Car {
     public static final String MOVEMENT_SYMBOL = "-";
     private static final int INITIAL_POSITION = 0;
-    private static final int MAX_NAME_LENGTH = 5;
     private static final int CRITERION_FOR_MOVING = 3;
 
-    private final String name;
+
+    private final CarName name;
     private int position;
 
 
     public Car(String name) {
-        validateName(name);
-        this.name = name;
+        this.name = new CarName(name);
         this.position = INITIAL_POSITION;
-    }
-
-    private void validateName(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("이름은 5자 이하여야 합니다.");
-        }
-        ;
     }
 
     public void move(int number) {
@@ -37,7 +29,7 @@ public class Car {
         return position;
     }
 
-    public String getName() {
+    public CarName getCarName() {
         return name;
     }
