+++ src/main/java/racingcar/domain/CarName.java
@@ -1,0 +1,22 @@
+package racingcar.domain;
+
+public class CarName {
+    private static final int MAX_NAME_LENGTH = 5;
+
+    private final String name;
+
+    public CarName(String name) {
+        validateName(name);
+        this.name = name;
+    }
+
+    private void validateName(String name) {
+        if (name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("이름은 5자 이하여야 합니다.");
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
