+++ src/main/java/racingcar/view/OutputView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Cars;
 
 import java.util.List;
